variable need to be int64

+++ supervised_learning/optimization/12-learning_rate_decay.py
@@ -6,7 +6,7 @@ import tensorflow as tf
 
 def learning_rate_decay(alpha, decay_rate, decay_step):
     """Creates an inverse time learning rate decay operation."""
-    global_step = tf.Variable(0, trainable=False)
+    global_step = tf.Variable(0, trainable=False, dtype=tf.int64)
 
     return tf.compat.v1.train.inverse_time_decay(
         alpha,
